feat(storybook): add addon-docs and modify dependencies
change awesome-typescript-loader for more trending ts-loader, upgrade all @storybook packages to
match addon-docs version, remove unnecssary react is, modify webpack and storybook config for
package upgrades

BREAKING CHANGE: replace awesome-typescript-loader for ts-loader, upgrade @storybook/*

fix #238
