+++ .storybook/webpack.config.js
@@ -1,5 +1,3 @@
-const { TsConfigPathsPlugin } = require('awesome-typescript-loader')
-const createCompiler = require('@storybook/addon-docs/mdx-compiler-plugin')
 const webpack = require('webpack')
 const path = require('path')
 
@@ -8,10 +6,10 @@ module.exports = async ({ config, mode }) => {
     test: /\.tsx?$/,
     use: [
       {
-        loader: require.resolve('awesome-typescript-loader'),
-        query: {
-          configFileName: path.resolve(__dirname, '../stories/tsconfig.json'),
-        },
+        loader: require.resolve('ts-loader'),
+        options: {
+          configFile: path.resolve(__dirname, '../stories/tsconfig.json'),
+        }
       },
       // Optional
       {
@@ -20,31 +18,12 @@ module.exports = async ({ config, mode }) => {
     ],
   })
   config.module.rules.push({
-    test: /\.(stories|story)\.mdx$/,
-    use: [
-      {
-        loader: 'babel-loader',
-        // may or may not need this line depending on your app's setup
-        options: {
-          plugins: ['@babel/plugin-transform-react-jsx'],
-        },
-      },
-      {
-        loader: '@mdx-js/loader',
-        options: {
-          compilers: [createCompiler({})],
-        },
-      },
-    ],
-  });
-  config.module.rules.push({
     test: /\.(stories|story)\.[tj]sx?$/,
     loader: require.resolve('@storybook/source-loader'),
     exclude: [/node_modules/],
     enforce: 'pre',
   });
   config.resolve.extensions = ['.ts', '.tsx', '.js', '.jsx']
-  config.resolve.plugins = [new TsConfigPathsPlugin({})]
   config.plugins.push(
     new webpack.DefinePlugin({
       __DEV__: process.env.NODE_ENV === 'development',
