+++ package.json
@@ -55,6 +55,7 @@
   },
   "devDependencies": {
     "@babel/core": "~7.7.7",
+    "@babel/helper-compilation-targets": "^7.8.3",
     "@commitlint/cli": "~8.3.3",
     "@commitlint/config-conventional": "~8.3.3",
     "@commitlint/prompt": "~8.3.3",
@@ -64,13 +65,13 @@
     "@semantic-release/github": "~5.5.3",
     "@semantic-release/npm": "~5.3.5",
     "@semantic-release/release-notes-generator": "~7.3.0",
-    "@storybook/addon-actions": "~5.2.6",
+    "@storybook/addon-actions": "~5.3.9",
     "@storybook/addon-docs": "^5.3.9",
-    "@storybook/addon-info": "~5.2.6",
-    "@storybook/addon-links": "~5.2.6",
-    "@storybook/addons": "~5.2.6",
-    "@storybook/react": "~5.2.6",
-    "@storybook/theming": "~5.2.6",
+    "@storybook/addon-info": "~5.3.9",
+    "@storybook/addon-links": "~5.3.9",
+    "@storybook/addons": "~5.3.9",
+    "@storybook/react": "~5.3.9",
+    "@storybook/theming": "~5.3.9",
     "@types/chart.js": "~2.9.7",
     "@types/enzyme": "~3.10.3",
     "@types/jest": "~24.0.25",
@@ -82,7 +83,6 @@
     "@types/tinymce": "~4.5.23",
     "@typescript-eslint/eslint-plugin": "~2.15.0",
     "@typescript-eslint/parser": "~2.15.0",
-    "awesome-typescript-loader": "~5.2.1",
     "babel-loader": "~8.0.6",
     "bootstrap": "~4.4.0",
     "commitizen": "~4.0.3",
@@ -110,9 +110,9 @@
     "react-docgen": "~5.1.0",
     "react-docgen-typescript-loader": "~3.6.0",
     "react-dom": "~16.12.0",
-    "react-is": "^16.12.0",
     "semantic-release": "~15.14.0",
     "sinon": "~8.0.1",
+    "ts-loader": "^6.2.1",
     "tsdx": "~0.12.0",
     "tslib": "~1.10.0",
     "typescript": "~3.7.4"
