+++ types/addon-docs-blocks.d.ts
@@ -1,0 +1,1 @@
+declare module '@storybook/addon-docs/blocks'
